Move meta tags management out to a container
Even though this sort of goes against what Helmet recommends, it's
easier and a little more performant with the use of selectors to manage
the generation of the tags. We tend to have to do a lot more client side
gathering of the tags (truncating text, combining images, etc) than we'd
like. Having them live in the container allows us to control how much
they're getting generated on re-renders and keeps them in one place.

Other things that happened here:

1. Added a bunch of selectors to cache the lookups
2. Moved some default tags that never change out to the html templates
3. Moved some string constants to locales/en.js file (more to here)
4. Cleaned up some of the parent containers and `params` properties
5. Added stubs for image and embed blocks
6. Added unit tests around the new selectors
7. Random cleanup
8. Removed the old helmet components