Updates scrolling to use react-router-scroll
Stores notification scroll positions in `sessionStorage` to set scroll
position when changing between notification types. This works on mobile
as well even though there isn't really a way to click on other types of
notifications without scrolling to the top, but figured if we ever made
the notification tabs on mobile sticky that we wouldn't need to touch
the scrolling behavior.

[Finishes #129568023](https://www.pivotaltracker.com/story/show/129568023)