Send users who have onboarding version 0 back
Our fix for onboarding version 1 to not have existing users go through
onboarding was to set their version to nil and new ones to '0'. We were
checking for onboarding versions matching and if they didn't we would
send them through it. Now initial users should be set to nil and thus
sent through version 2 of onboarding along with old initial users.